Restructure project so that all source code is in src folder
