+++ CMakeLists.txt
@@ -1,13 +1,17 @@
+# Root
 cmake_minimum_required(VERSION 3.0)
 project(FracTool)
 
+# Build config
 set(CMAKE_CXX_STANDARD 14)
 include_directories(include)
 
+# Packages
 find_package(OpenCV 4.5 REQUIRED)
 find_package(Boost 1.76 REQUIRED COMPONENTS log)
 
+# Debug mode settings
 add_compile_definitions($<$<CONFIG:Debug>:DEBUG=1>)
 
-add_subdirectory(apps)
+# Source directories
 add_subdirectory(src)
