+++ include/fractool/ftcore/config.hpp
