+++ include/fractool/ftcore/generate_mandelbrot.hpp
