+++ include/fractool/ftcore/map_color.hpp
