+++ include/fractool/ftcore/write_image.hpp
